chore: add `atuin-domain` crate for shared domain types
`atuin-common` had become a grab-bag of utilities mixed with Atuin's
core domain model. Carve the domain types out into a dedicated,
dependency-light `atuin-domain` crate shared across the client, daemon,
and server.

- Move `api` (HTTP request/response types) and `record` (the sync record
  model) from `atuin-common` into `atuin-domain`
- Inline the trivial `uuid_v7()` helper so `atuin-domain` depends on no
  other `atuin-*` crate
- Drop `atuin-common`'s now-unused `typed-builder`, `semver`, and `sqlx`
  dependencies
- Point all consumers at `atuin_domain::{api,record}`

History (still entangled with the client DB layer) is a natural next
candidate but is left for a follow-up.
