+++ crates/atuin-domain/src/api.rs
