+++ crates/atuin-dotfiles/src/store.rs
@@ -5,7 +5,7 @@ use atuin_client::record::sqlite_store::SqliteStore;
 // This will be noticeable similar to the kv store, though I expect the two shall diverge
 // While we will support a range of shell config, I'd rather have a larger number of small records
 // + stores, rather than one mega config store.
-use atuin_common::record::{DecryptedData, Host, HostId};
+use atuin_domain::record::{DecryptedData, Host, HostId};
 use atuin_common::utils::unquote;
 use eyre::{Result, bail, ensure, eyre};
 
@@ -238,7 +238,7 @@ impl AliasStore {
             .await?
             .map_or(0, |entry| entry.idx + 1);
 
-        let record = atuin_common::record::Record::builder()
+        let record = atuin_domain::record::Record::builder()
             .host(Host::new(self.host_id))
             .version(CONFIG_SHELL_ALIAS_VERSION.to_string())
             .tag(CONFIG_SHELL_ALIAS_TAG.to_string())
@@ -274,7 +274,7 @@ impl AliasStore {
             .await?
             .map_or(0, |entry| entry.idx + 1);
 
-        let record = atuin_common::record::Record::builder()
+        let record = atuin_domain::record::Record::builder()
             .host(Host::new(self.host_id))
             .version(CONFIG_SHELL_ALIAS_VERSION.to_string())
             .tag(CONFIG_SHELL_ALIAS_TAG.to_string())
@@ -386,7 +386,7 @@ mod tests {
             .await
             .unwrap();
         let key: [u8; 32] = XSalsa20Poly1305::generate_key(&mut OsRng).into();
-        let host_id = atuin_common::record::HostId(atuin_common::utils::uuid_v7());
+        let host_id = atuin_domain::record::HostId(atuin_common::utils::uuid_v7());
 
         let alias = AliasStore::new(store, host_id, key);
 
@@ -441,7 +441,7 @@ alias kgap='kubectl get pods --all-namespaces'
     #[tokio::test]
     async fn build_aliases_skips_corrupt_records() {
         use atuin_client::record::encryption::PASETO_V4;
-        use atuin_common::record::{DecryptedData, Host};
+        use atuin_domain::record::{DecryptedData, Host};
 
         use super::CONFIG_SHELL_ALIAS_TAG;
 
@@ -449,7 +449,7 @@ alias kgap='kubectl get pods --all-namespaces'
             .await
             .unwrap();
         let key: [u8; 32] = XSalsa20Poly1305::generate_key(&mut OsRng).into();
-        let host_id = atuin_common::record::HostId(atuin_common::utils::uuid_v7());
+        let host_id = atuin_domain::record::HostId(atuin_common::utils::uuid_v7());
 
         let alias = AliasStore::new(store.clone(), host_id, key);
 
@@ -458,7 +458,7 @@ alias kgap='kubectl get pods --all-namespaces'
         // a record in the alias tag encrypted with a different key - the store is corrupt,
         // or "mixed". it should be skipped, rather than breaking the build entirely.
         let corrupt_key: [u8; 32] = XSalsa20Poly1305::generate_key(&mut OsRng).into();
-        let corrupt = atuin_common::record::Record::builder()
+        let corrupt = atuin_domain::record::Record::builder()
             .host(Host::new(host_id))
             .version(CONFIG_SHELL_ALIAS_VERSION.to_string())
             .tag(CONFIG_SHELL_ALIAS_TAG.to_string())
