+++ crates/atuin-dotfiles/src/store/var.rs
@@ -5,7 +5,7 @@
 use std::collections::BTreeMap;
 
 use atuin_client::record::sqlite_store::SqliteStore;
-use atuin_common::record::{DecryptedData, Host, HostId};
+use atuin_domain::record::{DecryptedData, Host, HostId};
 use eyre::{Result, bail, ensure, eyre};
 
 use atuin_client::record::encryption::PASETO_V4;
@@ -285,7 +285,7 @@ impl VarStore {
             .await?
             .map_or(0, |entry| entry.idx + 1);
 
-        let record = atuin_common::record::Record::builder()
+        let record = atuin_domain::record::Record::builder()
             .host(Host::new(self.host_id))
             .version(DOTFILES_VAR_VERSION.to_string())
             .tag(DOTFILES_VAR_TAG.to_string())
@@ -321,7 +321,7 @@ impl VarStore {
             .await?
             .map_or(0, |entry| entry.idx + 1);
 
-        let record = atuin_common::record::Record::builder()
+        let record = atuin_domain::record::Record::builder()
             .host(Host::new(self.host_id))
             .version(DOTFILES_VAR_VERSION.to_string())
             .tag(DOTFILES_VAR_TAG.to_string())
@@ -470,7 +470,7 @@ mod tests {
             .await
             .unwrap();
         let key: [u8; 32] = XSalsa20Poly1305::generate_key(&mut OsRng).into();
-        let host_id = atuin_common::record::HostId(atuin_common::utils::uuid_v7());
+        let host_id = atuin_domain::record::HostId(atuin_common::utils::uuid_v7());
 
         let env = VarStore::new(store, host_id, key);
 
@@ -508,7 +508,7 @@ mod tests {
             .await
             .unwrap();
         let key: [u8; 32] = XSalsa20Poly1305::generate_key(&mut OsRng).into();
-        let host_id = atuin_common::record::HostId(atuin_common::utils::uuid_v7());
+        let host_id = atuin_domain::record::HostId(atuin_common::utils::uuid_v7());
 
         let env = VarStore::new(store, host_id, key);
