+++ crates/atuin-kv/Cargo.toml
@@ -16,6 +16,7 @@ readme.workspace = true
 [dependencies]
 atuin-client = { path = "../atuin-client", version = "18.18.1" }
 atuin-common = { path = "../atuin-common", version = "18.18.1" }
+atuin-domain = { path = "../atuin-domain", version = "18.18.1" }
 derive_more = { workspace = true }
 
 tracing = { workspace = true }
