+++ crates/atuin-scripts/src/store.rs
@@ -2,7 +2,7 @@ use eyre::{Result, eyre};
 
 use atuin_client::record::encryption::PASETO_V4;
 use atuin_client::record::sqlite_store::SqliteStore;
-use atuin_common::record::{Host, HostId, Record, RecordId, RecordIdx};
+use atuin_domain::record::{Host, HostId, Record, RecordId, RecordIdx};
 use record::ScriptRecord;
 use script::{SCRIPT_TAG, SCRIPT_VERSION, Script};
