+++ crates/atuin-scripts/src/store/record.rs
@@ -1,4 +1,4 @@
-use atuin_common::record::DecryptedData;
+use atuin_domain::record::DecryptedData;
 use eyre::{Result, eyre};
 use uuid::Uuid;
