+++ crates/atuin-server-postgres/Cargo.toml
@@ -11,6 +11,7 @@ repository = { workspace = true }
 
 [dependencies]
 atuin-common = { path = "../atuin-common", version = "18.18.1" }
+atuin-domain = { path = "../atuin-domain", version = "18.18.1" }
 atuin-server-database = { path = "../atuin-server-database", version = "18.18.1" }
 derive_more = { workspace = true }
