+++ crates/atuin-server/src/handlers/mod.rs
@@ -1,4 +1,4 @@
-use atuin_common::api::{ErrorResponse, IndexResponse};
+use atuin_domain::api::{ErrorResponse, IndexResponse};
 use atuin_server_database::Database;
 use axum::{Json, extract::State, http, response::IntoResponse};
