+++ crates/atuin/Cargo.toml
@@ -54,6 +54,7 @@ check-update = ["atuin-client/check-update"]
 atuin-ai = { path = "../atuin-ai", version = "18.18.1", optional = true, default-features = false }
 atuin-client = { path = "../atuin-client", version = "18.18.1", optional = true, default-features = false }
 atuin-common = { workspace = true, features = ["unicode"] }
+atuin-domain = { workspace = true }
 atuin-dotfiles = { workspace = true }
 atuin-history = { workspace = true }
 atuin-daemon = { path = "../atuin-daemon", version = "18.18.1", optional = true, default-features = false }
