+++ crates/tests-database/tests/db_story.rs
@@ -1,7 +1,5 @@
-use atuin_common::{
-    record::{EncryptedData, Host, HostId, Record, RecordIdx},
-    utils::{crypto_random_string, uuid_v7},
-};
+use atuin_common::utils::{crypto_random_string, uuid_v7};
+use atuin_domain::record::{EncryptedData, Host, HostId, Record, RecordIdx};
 use atuin_server_database::{
     Database, DbSettings, DbType,
     models::{NewSession, NewUser, User},
